test_make_credential_options
